requirements.txt
--find-links https://download.pytorch.org/whl/cu118

torch==2.5.1+cu118
torchvision==0.20.1+cu118
ultralytics>=8.0.0

matplotlib==3.8.2
numpy==1.26.4
opencv-python==4.9.0.80
Pillow==10.1.0
PyYAML==6.0.1
scipy==1.11.4
tqdm==4.66.1
tensorboard==2.15.1
seaborn==0.13.1
pandas==2.1.4

mss==9.0.1
pygame==2.5.2
pynput==1.7.6
pywin32==306
requests
wheel
termcolor==2.4.0
